* rocDecode/HW cap: Moved GpuVaContext class implementation back to VAAPI layer.
